should work
im not really sure, i did a lot of tests and i found out that in the y loop, we go throught all of the words including the file names so i just skiped them. Then we just take the index y+1 from what we read and it gives us the title of what we read.